+++ framework-generators/messaging-adapter-generator/src/test/java/uk/gov/justice/raml/jms/core/EventSourcesDestinationTypeFinderFactoryTest.java
@@ -1,0 +1,17 @@
+package uk.gov.justice.raml.jms.core;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
+
+import org.junit.jupiter.api.Test;
+
+class EventSourcesDestinationTypeFinderFactoryTest {
+
+    private final EventSourcesDestinationTypeFinderFactory factory = new EventSourcesDestinationTypeFinderFactory();
+
+    @Test
+    void shouldCreate() {
+        assertThat(factory.create(), is(notNullValue()));
+    }
+}