+++ CHANGELOG.md
@@ -4,7 +4,6 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
-
 ## [17.6.9] - 2024-10-11
 ### Fixed
 - Fixed spelling mistake in OversizeMessageGuard error message
@@ -45,6 +44,15 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 - Update framework libraries to 17.6.2 in order to:
   - Update jobstore to process tasks with higher priority first
   - Fix for Jackson single argument constructor issue inspired from  https://github.com/FasterXML/jackson-databind/issues/1498
+
+## [17.5.7] - 2024-08-12
+- For EVENT_PROCESSORs, derive the destinationType (queue|topic) from event_sources.location.jms_uri in event-sources.yaml when RAML is used for the EVENT_PROCESSOR
+
+## [17.5.6] - 2024-08-06
+- For EVENT_PROCESSORs, derive the destinationType (queue|topic) from event_sources.location.jms_uri in event-sources.yaml when YAML is  used for the EVENT_PROCESSOR
+
+## [17.5.5] - 2024-07-23
+- PEG-347: Jacksons SingleArgumentConstructor fix
 
 ## [17.6.1] - 2024-07-12
 ### Changed